requirements.txt
uvicorn[standard]
fastapi
pydantic
requests
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
timm==0.6.12
tqdm
pillow
pytorch_lightning==1.9.4
matplotlib
segmentation-models-pytorch
jupyter
albumentations
seaborn
japanize_matplotlib
Pillow>=7.1
matplotlib
pycocotools>=2.0.2
termcolor>=1.1
yacs>=0.1.8
tabulate
cloudpickle
tqdm>4.29.0
tensorboard
fvcore>=0.1.5,<0.1.6
iopath>=0.1.7,<0.1.10
dataclasses; python_version<'3.7'
omegaconf>=2.1,<2.4
hydra-core>=1.1
black
packaging
jinja2
python-multipart
